+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added Instructions input to AI Chat Component ([#87](https://github.com/architects-toolkit/SmartHopper/issues/87)).
+- Added systemPrompt parameter to WebChatUtils.ShowWebChatDialog
+
+### Changed
+
+- Several improvements to AI Chat Component:
+  - Updated WebChatDialog to use provided system prompt or fall back to default
+  - Improved default system prompt for AI Chat to focus on a Grasshopper assistant, including tool call examples
+- Modified manifest to reflect new context input feature
+- Code cleanup in AI Chat Component, WebChatDialog and WebChatUtils
+
 ## [0.3.3-alpha] - 2025-06-23
 
 ### WIP